Add RIM calculation
RIM depends not only on the provided images and memory layout, but
also on the content of DTB, which is dynamically generated during the startup.
The content of DTB higly depends on the kvmtool implementation and provided
options. For example, some virtual devices generate their portion of DTB in
the order of registration, which depends on the linking order (!).

To have exactly the same RIM, the measurement tool is based on the kvmtool
and TF-RMM source code. The tool re-uses the kvmtool code by conditionally
removing the code that launches VMs, keeping left the code responsible for
generating the DBT.